feat: per-type alert chimes + alert-mode approval
Three distinct device notifications, each a musical chime on the buddy speaker:
  - approve  (a gated tool needs permission) — C6, 400 BPM, x2
  - question (Claude calls AskUserQuestion)  — G6, 300 BPM, x3
  - complete (the turn finished)             — B6, 200 BPM, x4

Firmware gains a non-blocking beep sequencer (beepSeq: float freq, dur, BPM,
count, volume) driven by {"cmd":"beep","name":...}; buddyRequestBeep maps each
name to its pattern. The beep handler must sit before xferCommand() in data.h —
xferCommand swallows any unknown cmd when no transfer is active, which was
eating the beep command (silent, screen-only).

Approval gains an "alert" mode (now the default): the device shows the prompt
and chimes, but you decide in VS Code — the device shows "-> approve/answer in
editor" and its A/B buttons are inert (prompt.alert). button_approval also
supports true (device A/B decides) and false (display-only).

Defaults: button_approval="alert", approve_wait=0 (chime instantly; questions
and completion are always instant), busy_boost on. Daemon loop ticks at 0.1s
for low latency. Config keys and alert-sound tuning are documented in README.md
and bridge/NOTES.md.

+++ src/m5_compat.h
@@ -31,7 +31,7 @@ inline int chipTempC() { return (int)temperatureRead(); }
 inline void getAccel(float* ax, float* ay, float* az) {
   M5.Imu.getAccel(ax, ay, az);
 }
-inline void beep(uint16_t freq, uint16_t durMs) {
+inline void beep(float freq, uint16_t durMs) {
   M5.Speaker.tone(freq, durMs);
 }
 inline void powerOff() { M5.Power.powerOff(); }
